Moving routes into the app creator
